Fixed bug that lead to some external file lines (from censorship levels in mod) not loading
Default awk behaviour keeps files open and reads lines sequentially until the file is explicitly closed.
This meant that when subs in external files are called out of order (e.g. one half way into the file is called, then later, one at the beginning is called)
the desired sub is never found and that text never loaded.
So, to fix this, the file is closed after each sub has been fully processed.

Improved and added more comments.
Improved the method of determining where a sub starts and ends (counting open and close brackets for the subroutine until they cancel out to zero)